deleted old capture the flag folder
