+++ src/test/java/org/braekpo1nt/mctmanager/games/capturetheflag/CaptureTheFlagGameTest.java
@@ -1,0 +1,4 @@
+package org.braekpo1nt.mctmanager.games.capturetheflag;
+
+public class CaptureTheFlagGameTest {
+}
